Tag: V00-01-02 2011-05-23 Andy - suppress compilation warning about sscanf format (which could have   also been an error) in SciMDTestApp.cpp
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@1933 b967ad99-d558-0410-b138-e0f6c56caec7

+++ SciMD/trunk/app/SciMDTestApp.cpp
@@ -539,7 +539,7 @@ SciMDTestApp::cmd_set_run_param () throw (std::exception)
 
     } else if (type == "INT64") {
 
-        int64_t value_int ;
+        long long int value_int ;
         if (1 != sscanf(value.c_str(), "%lld", &value_int)) {
             MsgLogRoot (error, "parameter value is not of the claimed type") ;
             return 2 ;
@@ -550,7 +550,7 @@ SciMDTestApp::cmd_set_run_param () throw (std::exception)
             experiment,
             run,
             param,
-            value_int,
+            int64_t(value_int),
             "SciMDTestApp",
             m_update.value () > 0) ;
         m_connection->commitTransaction () ;
